requirements.txt
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiohttp-jinja2==1.5
aiohttp-session==2.12.0
aiokafka==0.8.0
aiosignal==1.3.1
altgraph==0.17.3
aniso8601==9.0.1
arcade==2.6.16
argcomplete==2.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
bcrypt==4.0.1
Beaker==1.10.0
beautifulsoup4==4.11.0
blivet==3.5.0
blivet-gui==2.4.0
Brlapi==0.8.4
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
cryptography==38.0.4
cupshelpers==1.0
dasbus==1.6
dbus-python==1.3.2
dependency-injector==4.40.0
Deprecated==1.2.13
distlib==0.3.6
distro==1.7.0
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.2.1
docopt==0.6.2
dulwich==0.20.50
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.8.2
fros==1.1
frozenlist==1.3.3
gpg==1.17.0
graphene==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.1
html5lib==1.1
humanize==3.13.1
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.1
kafka-python==2.0.2
keyring==23.11.0
langtable==0.0.61
libcomps==0.1.18
lockfile==0.12.2
lxml==4.9.1
Mako==1.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
meson==0.63.3
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.3
nautilus==0.5.2
nautilus-open-any-terminal==0.4.0
nftables==0.1
nose2==0.12.0
numpy==1.24.0
olefile==0.46
packaging==22.0
Paste==3.5.0
pastel==0.2.0
peewee==3.15.4
pexpect==4.8.0
pid==2.2.3
Pillow==9.1.1
pipreqs==0.4.11
pkginfo==1.9.2
platformdirs==2.6.0
pluggy==1.0.0
ply==3.11
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
productmd==1.33
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pwquality==1.4.4
pycairo==1.21.0
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.1
pyenchant==3.2.2
pyglet==2.0.dev23
Pygments==2.12.0
PyGObject==3.42.2
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.6.0
pykickstart==3.41
pylev==1.4.0
pymongo==4.3.2
pymunk==6.2.1
pynvim==0.4.3
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyparted==3.12.0
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.0
python-augeas==1.1.0
python-consul==1.1.0
python-dateutil==2.8.2
python-meh==0.50
pytiled-parser==2.2.0
pytz==2022.6
pyudev==0.23.2
pyxdg==0.27
PyYAML==6.0
redis==4.3.3
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-toolbelt==0.9.1
rpm==4.18.0
SecretStorage==3.3.3
selinux==3.4
sepolicy==3.4
setools==4.4.0
setproctitle==1.2.3
shellingham==1.5.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.4
soupsieve==2.3.2.post1
sqlcipher3==0.4.5
sqlparse==0.2.4
systemd-python==235
tcod==14.0.0
Tempita==0.5.2
tk==0.1.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
uvloop==0.17.0
validator-collection==1.5.0
virtualenv==20.17.1
webencodings==0.5.1
wrapt==1.14.0
yarg==0.1.9
yarl==1.8.2
zipp==3.11.0
